changed the advertisment functionality if gameover
createAd() function now only gives a one time increase of 10€ and a modal gets displayed if one of either machine is clicked (pizza machine/buyer machine) and the conditions are met (no pizza in storage and not enough money to buy new pizzas). before you had to click on buy pizza again in order for the ad button to display.